# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TradingBot API
  version: 0.1.0
  description: |
    Full API and WebSocket reference for the TradingBot Go backend.
    Base path for all REST endpoints: `/api/v1`
    WebSocket endpoints are documented under the `x-websocket` extension sections at the bottom.

    ## Authentication
    All protected endpoints require a Bearer JWT in the `Authorization` header.
    Tokens are obtained via `POST /api/v1/auth/login` or `POST /api/v1/auth/signup`.
    Access tokens expire in **15 minutes**. Use `POST /api/v1/auth/refresh` to rotate.

    ## Response Envelope
    Every response (success or error) is wrapped in a standard envelope:
    ```json
    { "data": {}, "meta": { "requestId": "", "serverTime": "", "version": "v1" } }
    { "error": { "code": "", "message": "", "details": {} }, "meta": { ... } }
    ```
servers:
  - url: http://localhost:8000
    description: Local development server
security:
  - BearerAuth: []

# ─────────────────────────────────────────────────────────────────────────────
# Security Schemes
# ─────────────────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        HS256-signed JWT. Obtain via /auth/login or /auth/signup.
        Include as: `Authorization: Bearer <token>`

  # ───────────────────────────────────────────────────────────────────────────
  # Shared Schemas
  # ───────────────────────────────────────────────────────────────────────────
  schemas:
    Meta:
      type: object
      properties:
        requestId:
          type: string
          example: req_1234567890
        serverTime:
          type: string
          format: date-time
          example: "2026-05-06T23:10:00+05:30"
        version:
          type: string
          example: v1
        pagination:
          $ref: '#/components/schemas/Pagination'
    Pagination:
      type: object
      properties:
        limit:
          type: integer
          example: 50
        cursor:
          type: string
        nextCursor:
          type: string
        hasMore:
          type: boolean
    APIError:
      type: object
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: One or more fields are invalid.
        details:
          nullable: true
    ErrorResponse:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/APIError'
        meta:
          $ref: '#/components/schemas/Meta'

    # ── Auth ──────────────────────────────────────────────────────────────────
    UserPayload:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: user@example.com
        userName:
          type: string
          example: Bhavik
        isActive:
          type: boolean
          example: true
    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserPayload'
        accessToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refreshToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        expiresIn:
          type: integer
          description: Access token TTL in seconds
          example: 900
    RefreshResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
          description: Rotated refresh token — old token is invalidated
        expiresIn:
          type: integer
          example: 900

    # ── Health ────────────────────────────────────────────────────────────────
...
    BrokerStatusData:
      type: object
      properties:
        broker:
          type: string
          example: ZERODHA
        connected:
          type: boolean
        brokerUserId:
          type: string
          example: AB1234
        accountName:
          type: string
          example: Zerodha Personal
        sessionExpiry:
          type: string
          format: date-time
          nullable: true
        lastSyncedAt:
          type: string
          format: date-time
          nullable: true
        tradingEnabled:
          type: boolean

    # ── Market ────────────────────────────────────────────────────────────────
    OHLC:
      type: object
      properties:
        open:
          type: number
          format: double
        high:
          type: number
          format: double
        low:
          type: number
          format: double
        close:
          type: number
          format: double
    QuoteItem:
      type: object
      properties:
        symbol:
          type: string
          example: NSE:RELIANCE
        instrumentToken:
          type: integer
          example: 738561
        lastPrice:
          type: number
          format: double
          example: 3000.5
        netChange:
          type: number
          format: double
          example: 15.25
        percentChange:
          type: number
          format: double
          example: 0.51
        volumeTraded:
          type: integer
          format: int64
          example: 123456
        ohlc:
          $ref: '#/components/schemas/OHLC'
        updatedAt:
          type: string
          format: date-time
    MarketOverviewItem:
      type: object
      properties:
        symbol:
          type: string
          example: NSE:RELIANCE
        lastPrice:
          type: number
          format: double
        percentChange:
          type: number
          format: double
        volume:
          type: integer
          format: int64
        bid:
          type: number
          format: double
        ask:
          type: number
          format: double
        updatedAt:
          type: string
          format: date-time
    MarketBreadth:
      type: object
      properties:
        advancers:
          type: integer
        decliners:
          type: integer
        unchanged:
          type: integer
    MarketOverviewData:
      type: object
      properties:
        indices:
          type: array
          items:
            $ref: '#/components/schemas/MarketOverviewItem'
        topGainers:
          type: array
          items:
            $ref: '#/components/schemas/MarketOverviewItem'
        topLosers:
          type: array
          items:
            $ref: '#/components/schemas/MarketOverviewItem'
        mostActiveByVolume:
          type: array
          items:
            $ref: '#/components/schemas/MarketOverviewItem'
        marketBreadth:
          $ref: '#/components/schemas/MarketBreadth'
        updatedAt:
          type: string
          format: date-time
...
  responses:
    Unauthorized:
      description: Missing or invalid Bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNAUTHORIZED
              message: invalid or expired token
            meta:
              requestId: req_123
              version: v1
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

# =============================================================================
# PATHS
# =============================================================================
paths:
...
  /api/v1/auth/signup:
    post:
      tags: [Auth]
      summary: Register a new user account
      operationId: signup
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  minLength: 8
                  example: StrongPass123!
                userName:
                  type: string
                  example: Bhavik
            example:
              email: user@example.com
              password: StrongPass123!
              userName: Bhavik
      responses:
        '201':
          description: User created — returns token pair
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/AuthResponse'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Authenticate and receive JWT token pair
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  example: StrongPass123!
            example:
              email: user@example.com
              password: StrongPass123!
      responses:
        '200':
          description: Login successful — returns token pair
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/AuthResponse'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Account inactive
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate refresh token and issue new access token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: New token pair issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/RefreshResponse'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: Invalidate refresh token and end session
      operationId: logout
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '204':
          description: Session ended — no body
        '400':
          $ref: '#/components/responses/BadRequest'

  # ── Profile ─────────────────────────────────────────────────────────────────
  /api/v1/me:
    get:
      tags: [Profile]
      summary: Get authenticated user profile
      operationId: getMe
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      email:
                        type: string
                        format: email
                        example: user@example.com
                      userName:
                        type: string
                        example: Bhavik
                      isActive:
                        type: boolean
                        example: true
                      createdAt:
                        type: string
                        format: date-time
                  meta:
                    $ref: '#/components/schemas/Meta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Profile]
      summary: Update display name
      operationId: patchMe
      description: Only `userName` can be updated here. Email and password changes require separate flows.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userName]
              properties:
                userName:
                  type: string
                  minLength: 1
                  maxLength: 100
                  example: Bhavik Updated
      responses:
        '200':
          description: Updated user profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                      email:
                        type: string
                      userName:
                        type: string
                      isActive:
                        type: boolean
                      createdAt:
                        type: string
                        format: date-time
                  meta:
                    $ref: '#/components/schemas/Meta'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  # ── Broker ──────────────────────────────────────────────────────────────────
  /api/v1/brokers/zerodha/status:
    get:
      tags: [Broker]
      summary: Get Zerodha broker connection status
      operationId: getBrokerStatus
      responses:
        '200':
          description: Broker status
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/BrokerStatusData'
                  meta:
                    $ref: '#/components/schemas/Meta'
              example:
                data:
                  broker: ZERODHA
                  connected: true
                  brokerUserId: AB1234
                  accountName: Zerodha Personal
                  sessionExpiry: "2026-05-07T06:00:00+05:30"
                  lastSyncedAt: "2026-05-06T23:10:00+05:30"
                  tradingEnabled: false
                meta:
                  requestId: req_123
                  version: v1
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/quotes:
    get:
      tags: [Market]
      summary: Batch quote lookup for multiple symbols
      operationId: getQuotes
      parameters:
        - name: symbols
          in: query
          required: true
          description: Comma-separated symbols e.g. NSE:RELIANCE,NSE:TCS
          schema:
            type: string
            example: NSE:RELIANCE,NSE:TCS
      responses:
        '200':
          description: Quote list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/QuoteItem'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/market/overview:
    get:
      tags: [Market]
      summary: Market snapshot — indices, top gainers/losers, breadth
      operationId: getMarketOverview
      responses:
        '200':
          description: Market overview snapshot
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/MarketOverviewData'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/market/heatmap:
    get:
      tags: [Market]
      summary: Current market heatmap snapshot for initial render
      operationId: getMarketHeatmap
      responses:
        '200':
          description: Heatmap snapshot
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MarketOverviewItem'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ── Watchlist ────────────────────────────────────────────────────────────────
...